confimation

+++ view/template/authors/all.view.php
@@ -17,12 +17,14 @@ ob_start();
                     <p class="card-text"><?= substr($author['biography'], 0 ,200) . '...' ?></p>
                     <a href="index.php?controller=authors&action=oneAuthor&id=<?= $author['id'] ?>"><i class="bi bi-eye-fill"></i></a>
                     <a href="index.php?controller=authors&action=updateAuthor&id=<?= $author['id'] ?>"><i class="bi bi-pencil-square"></i></a>
-                    <a href="index.php?controller=authors&action=deleteAuthor&id=<?= $author['id'] ?>"><i class="bi bi-trash-fill"></i></a>
+                    <a href="index.php?controller=authors&action=deleteAuthor&id=<?= $author['id'] ?>" onClick="confirm('Voulez-vous vraiment supprimer cet auteur ?')"><i class="bi bi-trash-fill"></i></a>
                 </div>
             </div>
         </div>
     <?php endforeach ?>
 </div>
+
+
 <?php
 $content = ob_get_clean();
